Task-14: Setting up code quality tools
