+++ .github/workflows/lint.yml
@@ -1,0 +1,28 @@
+name: linter
+
+on:
+  push:
+    branches: "**"
+  pull_request:
+    branches: [main]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+
+      - name: Setup Node
+        uses: actions/setup-node@v3
+        with:
+          node-version: "20"
+
+      - name: Install
+        run: npm install
+
+      - name: Prettier
+        run: npm run format
+
+      - name: Eslint
+        run: npm run lint
