+++ Task 1/index.js
@@ -1,2 +1,2 @@
-const sayHello = () => "Hello world!"
-console.log(sayHello())
+const sayHello = () => "Hello world!";
+console.log(sayHello());
