+++ app/utils/dateRangeException.js
@@ -1,0 +1,10 @@
+class DateRangeException extends Error {
+  constructor(message, statusCode) {
+    super(message);
+
+    this.statusCode = statusCode;
+    this.message = message;
+  }
+}
+
+module.exports = DateRangeException;
