+++ app/views/_layout.html
@@ -1,0 +1,10 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <title>Node Proxy Server</title>
+  </head>
+  <body>
+    {% block content %} {% endblock %}
+  </body>
+</html>
