+++ app/views/meteors.html
@@ -1,0 +1,22 @@
+{% extends '_layout.html' %} {% block content %}
+<h2>List of Meteors</h2>
+{% if meteors.length %}
+<ul>
+  {% for meteor in meteors %}
+  <li>
+    <strong>Name:</strong> {{meteor.name}}
+    <strong>Diameter:</strong> {{meteor.diameter_meters }} meters
+    <strong>Potentially Hazardous:</strong>
+    {% if meteor.is_potentially_hazardous_asteroid %} Yes {% else %} No {% endif
+    %}
+    <strong>Close Approach Date:</strong> {{ meteor.close_approach_date_full }}
+    <strong>Relative Velocity (km/s):</strong> {{ meteor.relative_velocity_kps
+    }}
+  </li>
+  {% endfor %}
+</ul>
+{% elif meteors.count %}
+<p>Number of meteorites: {{ meteors.count }}</p>
+{% else %}
+<p>No meteors found for the selected date range.</p>
+{% endif %} {% endblock %}
