+++ package.json
@@ -6,7 +6,12 @@
   "scripts": {
     "start": "cross-env NODE_TLS_REJECT_UNAUTHORIZED=0 node index.js",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon index.js"
+    "dev": "nodemon index.js",
+    "prepare": "husky install",
+    "pre-commit": "lint-staged",
+    "lint": "eslint .",
+    "lint:fix": "eslint --fix .",
+    "format": "prettier --write ."
   },
   "author": "",
   "license": "ISC",
@@ -18,6 +23,20 @@
     "nunjucks": "^3.2.4"
   },
   "devDependencies": {
-    "cross-env": "^7.0.3"
+    "@eslint/js": "^9.17.0",
+    "cross-env": "^7.0.3",
+    "eslint": "^9.17.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "globals": "^15.13.0",
+    "husky": "^8.0.0",
+    "lint-staged": "^15.2.11",
+    "prettier": "^3.4.2"
+  },
+  "lint-staged": {
+    "*.{js}": [
+      "prettier -w ./app",
+      "eslint ./app"
+    ]
   }
 }
