+++ prettierrc
@@ -1,0 +1,7 @@
+{
+  "semi": true,
+  "bracketSpacing": true,
+  "singleQuote": true,
+  "tabWidth": 2,
+  "trailingComma": "none"
+}
