--- routes/home-routes.js
@@ -1,8 +1,0 @@
-const express = require("express");
-const { getHomePage } = require("../controllers/home.controller.js");
-
-const router = express.Router();
-
-router.get("/", getHomePage);
-
-module.exports = router;
