--- views/_layout.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>Node Proxy Server</title>
-  </head>
-  <body>
-    {% block content %}
-    {% endblock %}
-  </body>
-</html>
